Remove old rudimentary code
