+++ exonum/src/api/mod.rs
@@ -29,9 +29,7 @@ use self::{
     node::{private::NodeInfo, public::ExplorerApi},
 };
 use crate::{
-    blockchain::{Blockchain, SharedNodeState},
-    crypto::PublicKey,
-    node::ApiSender,
+    api::node::SharedNodeState, blockchain::Blockchain, crypto::PublicKey, node::ApiSender,
 };
 
 pub mod backends;
